# cmd/api/feed.go
package main

import (
	"net/http"

	"github.com/Dylan-Oleary/go-social/internal/store"
)

// getUserFeedHandler godoc
//
//	@Summary		Fetches the user feed
//	@Description	Fetches the user feed
//	@Tags			feed
//	@Accept			json
//	@Produce		json
//	@Param			since	query		string	false	"Since"
//	@Param			until	query		string	false	"Until"
//	@Param			limit	query		int		false	"Limit"
//	@Param			offset	query		int		false	"Offset"
//	@Param			sort	query		string	false	"Sort"
//	@Param			tags	query		string	false	"Tags"
//	@Param			search	query		string	false	"Search"
//	@Success		200		{object}	[]store.PostWithMetadata
//	@Failure		400		{object}	error
//	@Failure		500		{object}	error
//	@Security		ApiKeyAuth
//	@Router			/users/feed [get]
func (app *application) getUserFeedHandler(w http.ResponseWriter, r *http.Request) {
	fq := store.PaginationFeedQuery{
		Limit:  20,
		Offset: 0,
		Sort:   "asc",
		Tags:   []string{},
	}

	fq, err := fq.Parse(r)
	if err != nil {
		app.badRequestError(w, err)
		return
	}

	if err := Validate.Struct(fq); err != nil {
		app.badRequestError(w, err)
		return
	}

	ctx := r.Context()

	// TODO: Auth User Id From Token
	feed, err := app.store.Posts.GetUserFeed(ctx, int64(43), fq)
	if err != nil {
		app.internalServerError(w, r, err)
		return
	}

	if err := app.jsonResponse(w, 200, feed); err != nil {
		app.internalServerError(w, r, err)
		return
	}
}